Add files via upload
Introduction

Welcome to the Customer Management System in Java! This project aims to provide a simple and efficient solution for managing customer data. Whether you are a small business owner or a developer looking for a customizable customer management tool, this Java-based system offers essential features for adding, searching, and displaying customer information.

    Key Features

    Add Customer Data: Easily add new customer details, including name, contact information, and any relevant notes.

    Search Data: Efficiently search for specific customers based on various parameters, ensuring quick access to relevant information.

    Display Data: View and manage customer data through a user-friendly interface, making it convenient to track and update customer records.



Usage
Please refer the snapshots.


Contact

Have questions or suggestions? Feel free to reach out to us via dmakwana1445@gmail.com  or open an issue in the repository.